fixed inconsistencies in getting exchange location
