+++ app/src/main/java/com/example/bookit/LocationActivity.java
@@ -2,8 +2,15 @@ package com.example.bookit;
 
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.FragmentTransaction;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
 
 public class LocationActivity extends AppCompatActivity {
     @Override
@@ -33,11 +40,25 @@ public class LocationActivity extends AppCompatActivity {
 
         else if (type == 2){
             GetLocationFragment fragment = new GetLocationFragment();
-            fragment.setArguments(bundle);
 
-            FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
-            transaction.add(R.id.common_container, fragment);
-            transaction.commit();
+            FirebaseFirestore db = FirebaseFirestore.getInstance();
+            DocumentReference docRef = db.collection("books").document(value);
+            docRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+                @Override
+                public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                    if (task.isSuccessful()) {
+                        DocumentSnapshot document = task.getResult();
+                        double latitude = Double.valueOf(document.get("latitude").toString());
+                        double longitude = Double.valueOf(document.get("longitude").toString());
+                        bundle.putDouble("latitude", latitude);
+                        bundle.putDouble("longitude", longitude);
+                        fragment.setArguments(bundle);
+
+                        FragmentTransaction transaction = getSupportFragmentManager().beginTransaction();
+                        transaction.add(R.id.common_container, fragment);
+                        transaction.commit();
+                    }}});
+
         }
 
     }
